ndsv5: add flag "--device-usb-combo"
This flag allows user to pass a bus-port-device number tuple to
ICEman, so that the assignment will not be disturb even in an
non-static environment.

Basically, we need this function to ensure better deployment process
for the BoardFarm project.  Currently, it is truely painful to add
even just one board into the database.  But that's another story.